+++ src/main/java/com/dailymap/dailymap/api/token/service/TokenService.java
@@ -6,6 +6,7 @@ import com.dailymap.dailymap.domain.jwt.service.TokenManager;
 import com.dailymap.dailymap.domain.member.model.Member;
 import com.dailymap.dailymap.domain.member.service.MemberService;
 import com.dailymap.dailymap.global.error.exception.BusinessException;
+import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,7 +17,7 @@ import static com.dailymap.dailymap.global.error.exception.ErrorCode.*;
 
 @Service
 @RequiredArgsConstructor
-public class AccessTokenReissueService {
+public class TokenService {
 
     private final MemberService memberService;
 
@@ -39,4 +40,13 @@ public class AccessTokenReissueService {
         return AccessTokenResponseDto.of(tokenDto);
     }
 
+    public boolean isRefreshTokenExpired(String authorization) {
+        String refreshToken = authorization.split(" ")[1];
+
+        Member findMember = memberService.findMemberByRefreshToken(refreshToken)
+            .orElseThrow(() -> new BusinessException(MEMBER_NOT_EXISTS_BY_REFRESH_TOKEN));
+
+        Date refreshTokenExpireTime = Timestamp.valueOf(findMember.getTokenExpirationTime());
+        return tokenManager.isTokenExpired(refreshTokenExpireTime);
+    }
 }
